requirements.txt
Django==2.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
mysqlclient==1.4.2
Pillow==6.1.0
pytz==2019.2
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
